third_party/blink/web_tests/printing/stretch-to-view-height-expected.txt

Test that the height of elements whose height depends on the viewport height is computed relative to the page height when printing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: number of pages is 1
PASS: number of pages is 1
PASS: number of pages is 1
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
